refactor : 패키지 구조 수정 및 dto 접근제한자 default로 변경
