+++ src/main/java/wooteco/subway/line/controller/LineController.java
@@ -1,8 +1,10 @@
-package wooteco.subway.line;
+package wooteco.subway.line.controller;
 
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import wooteco.subway.line.domain.Line;
+import wooteco.subway.line.LineService;
 
 import java.net.URI;
 import java.util.List;
