+++ src/main/java/wooteco/subway/line/controller/LineResponse.java
@@ -1,10 +1,10 @@
-package wooteco.subway.line;
+package wooteco.subway.line.controller;
 
-import wooteco.subway.station.StationResponse;
+import wooteco.subway.station.controller.StationResponse;
 
 import java.util.List;
 
-public class LineResponse {
+class LineResponse {
     private Long id;
     private String name;
     private String color;
