+++ src/main/java/wooteco/subway/line/domain/LineColor.java
@@ -1,4 +1,4 @@
-package wooteco.subway.line;
+package wooteco.subway.line.domain;
 
 import java.util.Objects;
