+++ src/main/java/wooteco/subway/station/dao/H2StationDao.java
@@ -1,10 +1,12 @@
-package wooteco.subway.station;
+package wooteco.subway.station.dao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import wooteco.subway.station.domain.Station;
+import wooteco.subway.station.domain.StationName;
 
 import java.sql.PreparedStatement;
 import java.util.List;
