+++ src/main/java/wooteco/subway/station/domain/StationName.java
@@ -1,4 +1,4 @@
-package wooteco.subway.station;
+package wooteco.subway.station.domain;
 
 import java.util.Objects;
 import java.util.regex.Pattern;
