+++ src/main/resources/application.properties
@@ -1,2 +1,3 @@
 spring.datasource.initialization-mode=always
 logging.config=classpath:logback.xml
+logging.level.root=debug
